Make algorithms take pointers instead of references to automata
This concerns various algorithms.
The algorithms required the caller to keep the references alive, which is unpleasant.
This is now still possible with old-fashioned pointers, but smart pointers are recommended.
The Python wrapper is updated to always hold automata by std::shared_ptr.

This relies on the new version of the Utility library, which has unique_ptr.hpp.